# Add safe refs for no-op 'map' for List and Map