--- a/__tests__/Map.ts
+++ b/__tests__/Map.ts
@@ -255,6 +255,12 @@ describe('Map', () => {
     expect(r.toObject()).toEqual({ A: 'a', B: 'b', C: 'c' });
   });
 
+  it('maps no-ops return the same reference', () => {
+    const m = Map({ a: 'a', b: 'b', c: 'c' });
+    const r = m.map(value => value);
+    expect(r).toBe(m);
+  });
+
   it('filters values', () => {
     const m = Map({ a: 1, b: 2, c: 3, d: 4, e: 5, f: 6 });
     const r = m.filter(value => value % 2 === 1);